refactor: separate runtime into different modules for better directory structure
