+++ cmd/sealos/cmd/cert.go
@@ -24,7 +24,7 @@ import (
 	"github.com/labring/sealos/pkg/apply/processor"
 	"github.com/labring/sealos/pkg/clusterfile"
 	"github.com/labring/sealos/pkg/constants"
-	"github.com/labring/sealos/pkg/runtime"
+	"github.com/labring/sealos/pkg/runtime/kubernetes"
 )
 
 var altNames string
@@ -58,7 +58,7 @@ func newCertCmd() *cobra.Command {
 			if err = cf.Process(); err != nil {
 				return err
 			}
-			r, err := runtime.NewDefaultRuntime(cluster, cf.GetKubeadmConfig())
+			r, err := kubernetes.NewDefaultRuntime(cluster, cf.GetKubeadmConfig())
 			if err != nil {
 				return fmt.Errorf("get default runtime failed, %v", err)
 			}
