+++ pkg/runtime/kubernetes/runtime_getter.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package runtime
+package kubernetes
 
 import (
 	"context"
@@ -148,11 +148,13 @@ func (k *KubeadmRuntime) execIPVSPod(ip string, masters []string) error {
 func (k *KubeadmRuntime) execToken(ip, certificateKey string) (string, error) {
 	return k.getRemoteInterface().Token(ip, k.initMasterKubeadmConfigFile(), certificateKey)
 }
+
 func (k *KubeadmRuntime) execHostname(ip string) (string, error) {
 	hostname, err := k.getRemoteInterface().Hostname(ip)
 	// tips: if hostname is upper,kubelet init master0 is not allowed to modify node
 	return strings.ToLower(hostname), err
 }
+
 func (k *KubeadmRuntime) execHostsAppend(ip, host, domain string) error {
 	return k.getRemoteInterface().HostsAdd(ip, iputils.GetHostIP(host), domain)
 }
@@ -190,7 +192,7 @@ func (k *KubeadmRuntime) getImageLabels() map[string]string {
 }
 
 func (k *KubeadmRuntime) getSSHInterface() ssh.Interface {
-	return k.ClusterClient
+	return k.clusterClient
 }
 
 func (k *KubeadmRuntime) getEnvInterface() env.Interface {
@@ -198,30 +200,29 @@ func (k *KubeadmRuntime) getEnvInterface() env.Interface {
 }
 
 func (k *KubeadmRuntime) getRemoteInterface() remote.Interface {
-	return remote.New(k.getClusterName(), k.getSSHInterface())
+	return k.remoteUtil
 }
 
 func (k *KubeadmRuntime) getContentData() constants.Data {
-	return constants.NewData(k.getClusterName())
+	return k.pathResolver
 }
 
-func (k *KubeadmRuntime) getKubeInterface() kubernetes.Client {
+func (k *KubeadmRuntime) getKubeInterface() (kubernetes.Client, error) {
 	if k.cli != nil {
-		return k.cli
+		return k.cli, nil
 	}
 	cli, err := kubernetes.NewKubernetesClient(k.getContentData().AdminFile(), k.getMaster0IPAPIServer())
 	if err != nil {
-		logger.Error("get kube client failed %v", err)
-		return nil
+		return nil, err
 	}
 	k.cli = cli
-	return cli
+	return cli, nil
 }
 
-func (k *KubeadmRuntime) getKubeExpansion() kubernetes.Expansion {
-	ki := k.getKubeInterface()
-	if ki != nil {
-		return kubernetes.NewKubeExpansion(ki.Kubernetes())
+func (k *KubeadmRuntime) getKubeExpansion() (kubernetes.Expansion, error) {
+	ki, err := k.getKubeInterface()
+	if err != nil {
+		return nil, err
 	}
-	return nil
+	return kubernetes.NewKubeExpansion(ki.Kubernetes()), nil
 }
