+++ pkg/runtime/types/consts.go
@@ -1,0 +1,11 @@
+package types
+
+const (
+	KubeadmV1beta1 = "kubeadm.k8s.io/v1beta1"
+	KubeadmV1beta2 = "kubeadm.k8s.io/v1beta2"
+	KubeadmV1beta3 = "kubeadm.k8s.io/v1beta3"
+
+	DefaultAPIServerDomain = "apiserver.cluster.local"
+	DefaultDNSDomain       = "cluster.local"
+	DefaultAPIServerPort   = 6443
+)
